Made the window bigger (updated background image to come), removed as many hard coded variables as necessary and replaced them with variables, added more comments

+++ src/crate.h
@@ -28,7 +28,7 @@ private:
   
   float x;
   float y;
-  int crateSize;
+  short crateSize;
   
 };
